fix: tests and docker-compose

+++ src/app/tests/conftest.py
@@ -6,14 +6,15 @@ from typing import AsyncGenerator
 import asyncio
 
 from main import app
-from main import configs
+from config import configs
 from app import engine, Base
 
 @pytest.fixture(scope="session", autouse=True)
 async def setup_db():
     async with engine.begin() as conn:
-        await conn.run_sync(Base.metadata.drop_all)
-        await conn.run_sync(Base.metadata.create_all)
+        if configs.mode =='test':
+            await conn.run_sync(Base.metadata.drop_all)
+            await conn.run_sync(Base.metadata.create_all)
 
 @pytest.fixture(scope="session")
 def user():
